requirements.txt
openai==1.0.0
mysql-connector-python==8.1.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
google-api-python-client==2.93.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
serpapi==1.0.0
